Allows labels to be converted into attributes
This change makes it possible to use labels as attributes by adding a trait bound
`Into` to the `Label` type parameter in various parts of the codebase, including
`PointStore`, `NodeStore`, `RCFBuilder`, and related structs and traits.

This allows the system to automatically convert labels into attributes when needed, which can simplify
code and improve performance. This change also includes an implementation for MultiTRCFLabel to enable
it to be converted into a u64, enabling labels to be used as attributes in MultiTRCF.